+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `onBecameBlurred` and `onBecameFocused` are always invoked synchonously with focus change and not on componentDidUpdate
 ### Added
-- `setFocus` and `navigateByDirection` accept an event object, this object is passed back on `onBecameBlurred` and `onBecameFocused` callbacks
+- `setFocus` and `navigateByDirection` accept an details object, this object is passed back on `onBecameBlurred` and `onBecameFocused` callbacks
 
 ## [2.10.0]
 ### Changed